pyproject.toml
[project]
name = "adk-secure-sessions"
version = "1.0.3"
description = "Encrypted session persistence for Google ADK agents"
readme = "README.md"
license = { text = "Apache-2.0" }
authors = [
    { name = "Alberto-Codes", email = "alberto.codes.dev@gmail.com" }
]
maintainers = [
    { name = "Alberto-Codes", email = "alberto.codes.dev@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.22.0",
    "cryptography>=44.0.0",
]
keywords = [
    "adk",
    "agents",
    "ai",
    "compliance",
    "encrypted sessions",
    "encryption",
    "google adk security",
    "hipaa",
    "security",
    "session encryption",
    "sessions",
    "sqlite",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Healthcare Industry",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Security :: Cryptography",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]

[project.optional-dependencies]
postgres = []  # Phase 3: PostgreSQL persistence backend

[project.urls]
Homepage = "https://github.com/Alberto-Codes/adk-secure-sessions"
Documentation = "https://alberto-codes.github.io/adk-secure-sessions/"
Repository = "https://github.com/Alberto-Codes/adk-secure-sessions"
Issues = "https://github.com/Alberto-Codes/adk-secure-sessions/issues"
Changelog = "https://github.com/Alberto-Codes/adk-secure-sessions/releases"

[build-system]
requires = ["uv_build>=0.9.28,<0.11.0"]
build-backend = "uv_build"

# TestPyPI publishing index (for `uv publish --index testpypi`)
[[tool.uv.index]]
name = "testpypi"
url = "https://test.pypi.org/simple/"
publish-url = "https://test.pypi.org/legacy/"
explicit = true

[tool.ruff]
line-length = 88
target-version = "py312"
exclude = [".uv-cache", ".venv", "__pycache__", "*.egg-info", "specs"]

[tool.ruff.lint]
extend-select = ["E501", "I", "D"]

[tool.ruff.lint.pycodestyle]
max-line-length = 100

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
"tests/*.py" = ["S101"]

[tool.ruff.lint.isort]
known-first-party = ["adk_secure_sessions"]

[tool.ruff.format]
quote-style = "double"
docstring-code-format = true
docstring-code-line-length = "dynamic"

[tool.docvet]
fail-on = ["enrichment", "freshness", "coverage", "griffe", "presence"]

[tool.docvet.presence]
min-coverage = 100.0

[tool.pytest.ini_options]
markers = [
    "unit: Unit tests - fast, isolated",
    "integration: Integration tests with real databases",
    "slow: Slow-running tests",
    "benchmark: Performance benchmark tests",
]
addopts = "-m 'not benchmark'"
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
filterwarnings = [
    "error",
    # aiosqlite.Connection.close() does not join _connection_worker_thread — the
    # thread may call call_soon_threadsafe() after pytest-asyncio closes the event
    # loop, raising RuntimeError("Event loop is closed"). This is an inherent race
    # in aiosqlite's architecture (no thread.join() in close()), not a test defect.
    "ignore:Exception in thread.*_connection_worker_thread.*:pytest.PytestUnhandledThreadExceptionWarning",
    # google-adk RunConfig accesses a deprecated pydantic field
    # (save_input_blobs_as_artifacts) internally — upstream issue, not our defect.
    # google-adk RunConfig.save_input_blobs_as_artifacts deprecated pydantic field;
    # narrowed to google.adk.runners to surface deprecations from other ADK submodules.
    "ignore:deprecated:DeprecationWarning:google\\.adk\\.runners",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
    "pytest-timeout>=2.4.0",
    "ruff>=0.13.0",
    "ty>=0.0.1a20",
    "google-adk>=1.23.0",
    "griffe>=1.15.0",
    "mkdocs-gen-files>=0.6.0",
    "mkdocs-material>=9.7.1",
    "mkdocstrings>=1.0.2",
    "mkdocstrings-python>=2.0.1",
    "mkdocs-literate-nav>=0.6.2",
    "mkdocs-section-index>=0.3.10",
    "mkdocs-git-revision-date-localized-plugin>=1.5.1",
    "mkdocs-minify-plugin>=0.8.0",
    "mkdocs-glightbox>=0.5.2",
    "mkdocs-macros-plugin>=1.5.0",
    "pillow>=12.1.0",
    "cairosvg>=2.8.2",
    "griffe-warnings-deprecated>=1.1.0",
    "griffe-inherited-docstrings>=1.1.2",
    "docvet>=1.7.0",
    "pip-audit>=2.9.0",
    "uv-secure>=0.17.0",
]
